NussknackerAppFactory vs AkkaHttpBasedRouteProvider: easier responsibility split
